+++ src/MetaheuristicsPlatform/Algorithms/HyperHeuristics/TabuSearchHyperHeuristic/TabuSearchHyperHeuristicOptimizerReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.HyperHeuristics.TabuSearchHyperHeuristic;
+
+public static class TabuSearchHyperHeuristicOptimizerReferences
+{
+    public static ScientificReference Primary { get; } =
+        new(
+            "Burke, Kendall & Soubeiga",
+            2003,
+            "A Tabu-Search Hyperheuristic for Timetabling and Rostering",
+            "Journal of Heuristics 9(6), 451-470",
+            "10.1023/B:HEUR.0000012446.94732.B6");
+}